Refactor environment variables, composer dependencies, and controller code
- Updated `APP_ENV` in `.env` from `dev` to `prod`
- Added `symfony/mime` as a dependency in `composer.json`
- Updated `composer.lock` with new hash and added `symfony/mime` entry
- Removed `Procfile`
- Added `images_directory` parameter in `config/services.yaml`
- Modified `ProductCrudController` to display the user's email instead
of the object
- Modified `ProductController` to include unused form creation and
retrieval of the currently logged in user

These changes aim to refine the project's configuration and ensure
accurate data representation. The update of the environment variable
aligns the application with the production environment, while the
addition of `symfony/mime` introduces MIME message manipulation
functionality. The removal of `Procfile` indicates a shift in deployment
strategy. The modifications to the controllers improve data
representation and code efficiency.

No issue references at the moment.